+++ agent.log
@@ -1,0 +1,2 @@
+2025-07-08 12:38:55,404 - main - INFO - Starting Autonomous AI Agent System
+2025-07-08 12:38:55,404 - main - INFO - Starting web interface on 0.0.0.0:8000
